requirements.txt
torch
hydra-core
omegaconf
einops
torchvision
numpy
pandas
tqdm
pillow
